Aligned storage of 'path' type.
Related change is that container of namespace declarations now points to offsets in the binary parser output as opposed to pointers. The pointers broke the parser as soon as reallocation took place due to increased volume of the aligned paths inside XNT. It was fixed in this commit to keep regression tests working.

+++ src/xnode--0.7.1.sql
@@ -68,7 +68,8 @@ CREATE TYPE @extschema@.path (
 	internallength = variable,
 	input = @extschema@.path_in,
 	output = @extschema@.path_out,
-	alignment = int,
+	-- 8-aligned because float8 is used to represent numbers
+	alignment = double,
 	storage = plain
 );
 
@@ -252,7 +253,8 @@ CREATE TYPE xnt (
 	internallength = variable,
 	input = xnt_in,
 	output = xnt_out,
-	alignment = int,
+	-- 8-aligned because it may contain XPath expressions
+	alignment = double,
 	storage = extended
 );
